refactor(cable): move Noise protocol onto the CableDataChannel abstraction
Lifts the Noise handshake, CableTunnelMessage framing, padding, the
connection loop and post-handshake/update parsing out of tunnel.rs into
a new protocol.rs, generic over a CableDataChannel instead of a
WebSocketStream. tunnel.rs keeps the WebSocket connect and tunnel-domain
decoding. connection_stages.rs carries a Box<dyn CableDataChannel>.

caBLE over WebSocket behaves identically; the abstraction lets a future
data channel (BLE L2CAP) plug in.
